Meeting 3-12-2019
For this meeting, we realized that we haven't been working as a team. We've all been doing our own things and we realize the importance of weekly team meetings to keep each other on track. Therefore, we will be meeting again this week on Thursday at 9 PM through Skype. In this meeting, we will share our own technologies and familiarize each member. That way, we can all work together on one technology as opposed to splitting the group into teams. The meeting minutes documentation provides more insight on what the meeting focused on and our goals for next meeting.