I added functionality to the floor options after clicking Capen Library
In CapenFloorOptions.java, I changed the adapter used to add functionality the Floor 1, Floor 2, and Floor 3 buttons in the ListView. I added onItemClickListeners to each of the items in the list. So far, they do not bring the user to a new activity with the floor plan, but we plan to add that functionality soon.