I created a new activity for Capen's first floor floorplan.
The file is called CapenFloor1Plan.java. I basically used a dummy image in place of the floor plan because we need to convert our floorplans to jpgs before we used them (they are PDFs as of now). This means that user story 1 is almost completed! Later in the week I will replace the picture with the actual floor plan and implement the rest of the floor options.