Added new test files for CapenFloorOptions and Capen1FloorPlan.
I started 3 tests for CapenFloorOptions and created the test file only for Capen1FloorPlan. The three tests I implemented are in CapenFloorOptionsTest.java located in AndroidTest. They test that the options in our ListView are the correct strings. Next time, I will test that when clicked, the options brings the user to the corresponding new activity and floor plan.