Added more tests to CapenFloorOptions.java.
I added a few more tests to CapenFloorOptions.java that tests if the listview is null or not, and if the Floor 2 button brings the user to the CapenFloor1Plan.java activity. I have also added some comments on top of my tests to state what they are testing for.