Created RegisterTest.java test file
I created the test file along with the before and after functions. In addition to them, I started on a test that checks that the register button when clicked, leads the user to the home activity (home page). We will soon incoporate Aditya's tests in the test file as well.