fixed failing tests in RegisterTest.java.
removed a test that we found was failing. It turns out what we were testing for was part of a different activity. The test we removed was for the UBLogo. Additionally we fixed a test that tests when the user presses the register button without filling in all the necessary fields that the home page does not show up.