added test to RegisterTest2.java.
I added a test that uses espresso to fill in all the fields required before registering such as email, password, etc. Afterwards, I automate AS to click on accept terms and agreements and click the register button. The test checks that after this action, the Login page should appear on the screen.