Created LocationActivity.java file
Created a new activity that will allow the user to determine their current location or search their current location from a list of options. I have created the corresponding xml file and menu xml file as well. So far, I have created the search options in a list view in the java file but will need to complete the xml file.