completed layout of LocationActivity.java
I implemented the button listeners for all the options in the list view. I have also implemented a clickable textview, that allows the user to get their current location based on Wifi. Additionally, I have also added the hint for the search bar and changed the text colors/layout in both LocationActivity.java and DestinationActivity.java.