Added the functionality of uploading an image
I worked on the activity which is led to after the user clicks 'Saved Paths' on the home screen's menu bar. It allows the user to select an image from his or her photo gallery and upload that to our firebase directory. I will continue working on the ability to retrieve images and show them in a gallery in another activity.