Added three tests to ProfileActivityTest.java
The first test checks that the profile avatar image view is using the correct imageView from the corresponding xml file. This is also done for the profile name text view using the assertEquals statement. The second test is ensuring the profile name text view appears on the screen when the activity is run.