Added three tests for the profile options list.
The three tests are very similar- they all test that the options appearing on the screen for the user profile, when clicked, brings the user to the right activity afterwards. I created three monitors- one for the messages, public paths, and settings activities. Then I check that when an option is clicked, the correct monitor appears on the screen.