Added three tests to SavedPathsTest.java.
The first test ensures the activity itself is not null when the application is run. the second test checks that the image view for the chosen path image appears on the screen. The last test checks that the image view being used is the correct image view from the corresponding xml file.