Implemented instant messaging
After finishing my chat message activity and corresponding xml file, I worked on the actual messaging activity, which allows the user to send a message instantly. This message will then be saved to firebase with the user's account, including the time stamp and date sent with the message. The user's email is also shown.