Created JUnit test file for the MessagesActivity.
I created the test file for our MessagesActivity.java. So far, I have created the activity test rule for the file along with the before and after methods. Additionally, I have also created the first test, which checks that the messages activity is not null when the application is ran.