Added three tests to MessagesActivityTest.java
The first test checks that when a user sends a message, the message sent is not null. The second test checks that when a user sends a message, the message user displayed is not null. The third test checks that when a user sends a message, the message time is not null.